+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 * add `MarginPoolCommand::Derive` subcommand.
+* add `MarginPoolCommand::Get` to retrieve specific or all margin pool accounts.
 
 ## [0.1.8] - 2022-06-05